+++ packages/feat-db/src/index.js
@@ -1,3 +1,4 @@
 export { default as dbImporter } from './dbImporter';
 export { default as supabase } from './supabase';
+export { default as postgres } from './postgres';
 export * from './excel';