+++ packages/jacaranda/src/helpers/connect.js
@@ -1,0 +1,7 @@
+// connect two koa middleware functions
+
+export default function connect(...middlewares) {
+    return async (ctx, next) => middlewares.reverse().reduce((prev, curr) => {
+        return () => curr(ctx, prev);
+    }, next)();
+}