+++ packages/jacaranda/test/controllers/actions/simple.js
@@ -6,4 +6,8 @@ export default {
     index2_: async (ctx) => {
         ctx.body = 'Hello, Jacaranda! 2';
     },
+
+    mc_: async (ctx) => {
+        ctx.body = `Hello, Jacaranda! ${ctx.mc}`;
+    },
 };